raid-app/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
image==1.5.27
imageio==2.9.0
numpy==1.16.2
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
Pillow==6.0.0
pydicom==1.3.0
torch==1.5.1+cpu
torchvision==0.6.1+cpu
